Convert getChangelog to pure functions
